move QtumPrediction Store
