+++ src/scenes/QtumPrediction/store.js
@@ -1,8 +1,8 @@
 import { observable, action, runInAction, reaction } from 'mobx';
 import _ from 'lodash';
 import { Token, OracleStatus, Routes } from 'constants';
-import { queryAllOracles } from '../network/graphQuery';
-import Oracle from './models/Oracle';
+import { queryAllOracles } from '../../network/graphQuery';
+import Oracle from '../../stores/models/Oracle';
 
 const INIT_VALUES = {
   loaded: false, // loading state?
